clean up tap setup, global logger
