+++ e2e/main_other_test.go
@@ -1,0 +1,9 @@
+//go:build e2e && !linux
+
+package e2e
+
+import "fmt"
+
+func mainSetup() error {
+	return fmt.Errorf("e2e tests are not supported on non-Linux platforms")
+}
